Synchronize Zabbix templates for Yoda
The template was somewhat out of sync.

Add the current versions. They are optimized for a "two server" setup of
Yoda, i.e. an "icat" that contains the iCAT enabled server, database and
webserver (portal and webdav) and a "resource" that only contains a non
iCAT enabled resource server.

The templates here are committed "as-is". They contain naming for "Yoda
1.4" but are used with Yoda 1.8 . Although some metrics probably do not
function anymore as originally intended.

The existing template (in this repository) is kept for now, to make
comparisons slightly easier.